Keep code location creation code only where its needed and some maintainability refactoring
