+++ src/main/java/com/blackduck/integration/detect/lifecycle/run/operation/OperationRunner.java
@@ -570,9 +570,7 @@ public class OperationRunner {
 
             Response response = blackDuckApiClient.execute(buildBlackDuckResponseRequest);
             String contentString = response.getContentString();
-            ScanCreationResponse scanCreationResponse = gson.fromJson(contentString, ScanCreationResponse.class);
-            
-            return scanCreationResponse;
+            return gson.fromJson(contentString, ScanCreationResponse.class);
         });
     }
